refactor(chat): split oversized chat files under 700 loc
